Create Railway.java
53.	Write an application for booking railway ticket reservation system. The application should have four functionalities.
1.	Book
2.	Cancel
3.	Print booked tickets (details with summary)
4.	Print available tickets (details with summary)
Conditions for booking:
There are a total of 63 berths for 63 confirmed tickets, 9 berths for 18 RAC tickets and 10 tickets in waiting-list. If the waiting-list ticket count goes above 10, print as ‘No tickets available’. The following passenger details should be obtained from the user.
1.	Name
2.	Age
3.	Gender
4.	Berth Preference
The tickets should not be allocated for children below age 5.But, their details should be stored. Lower berth should be allocated for persons whose age is above 60 and ladies with children if available. Side-lower berths should be allocated for RAC passengers.
Conditions for cancelling:
Whenever a ticket is cancelled, a ticket from RAC should be confirmed and a waiting-list ticket should move to RAC.
Conditions for printing booked tickets:
Print all the tickets that are filled along with the passenger details and at the end, print the total number of tickets that are filled.
Conditions for printing available tickets:
Print all the tickets that are unoccupied and at the end, print the total number of tickets that are unoccupied.